+++ compression_entropy.py
@@ -1,10 +1,11 @@
 import os
 import random
+from collections import Counter
 
 import numpy as np
 
 import data
-from word_pasting import get_char_distribution, select_samples, parse_mismatcher_lines
+from word_pasting import get_word_mismatches
 
 
 def read_selected_verses(filename: str,
@@ -102,3 +103,79 @@ def get_entropies(sample_verses: list,
     version_entropy = {version: get_entropy(mismatches)
                        for version, mismatches in version_mismatches.items()}
     return version_entropy
+
+
+def get_char_distribution(text: str) -> dict:
+    return Counter(text)
+
+
+def select_samples(sample_sequences: dict, chosen_sample_ids: list, truncate_samples: bool) -> dict:
+    if not chosen_sample_ids:
+        chosen_sample_ids = list(sample_sequences.keys())
+    lengths = {sample_id: get_text_length(sample_sequences[sample_id])
+               for sample_id in chosen_sample_ids
+               if sample_id in sample_sequences}
+    if len(lengths) == 0:
+        return {}
+    minimum_length = min(lengths.values())
+    differences = {sample_id: length - minimum_length for sample_id, length in lengths.items()}
+    full_samples = {sample_id: sample_sequences[sample_id]
+                    for sample_id in chosen_sample_ids
+                    if sample_id in sample_sequences}
+    if truncate_samples:
+        return {sample_id: truncate(sequences, differences[sample_id]) for sample_id, sequences in full_samples.items()}
+    return full_samples
+
+
+def parse_mismatcher_lines(lines: list) -> list:
+    return [int(line.split('\t')[-1].strip()) for line in lines if line != '\n']
+
+
+def get_text_length(sequences: list) -> int:
+    text = join_verses(sequences, insert_spaces=True)
+    return len(text)
+
+
+def truncate(sequences: list, surplus: int) -> list:
+    """
+    Truncate a sample by removing the number of characters in the surplus
+    :param sequences: a sample represented as a list of sequences, each of which is a list of tokens
+    :param surplus: the surplus that should be removed
+    :return: a new list of sequences, the length of which is reduced by the surplus
+    """
+    orig_len = get_text_length(sequences)
+    desired_length = orig_len - surplus
+    output = [seq.copy() for seq in sequences]
+    while get_text_length(output) > desired_length:
+        if len(output[-1]) > 1:
+            output[-1].pop()
+        else:
+            output.pop()
+    return output
+
+
+def join_verses(verse_tokens: list, insert_spaces: bool) -> str:
+    """
+    Join the verses contained in a list of lists of tokens
+    :param verse_tokens: the list of verses, each of which is a list of tokens; order matters
+    :param insert_spaces: whether we want to insert spaces between the tokens
+    :return: the concatenated string consisting of all the tokens in the original order
+    """
+    sep = ' ' if insert_spaces else ''
+    return sep.join([sep.join(ell) for ell in verse_tokens])
+
+
+def get_entropies_per_word(sample_verses: list,
+                           base_filename: str,
+                           remove_mismatcher_files: bool,
+                           mismatcher_path: str) -> float:
+    """
+    Get three entropies for a given sample of verses
+    :param sample_verses: the (ordered) pre-processed verses contained in the original sample
+    :param base_filename: the base filename to be used for the output
+    :param remove_mismatcher_files: whether to delete the mismatcher files after processing
+    :param mismatcher_path: path to mismatcher jar
+    :return: the entropies for the given sample (e.g., chapter)
+    """
+    # Compute the entropy
+    return get_entropy(get_word_mismatches(sample_verses, base_filename, remove_mismatcher_files, mismatcher_path))
