+++ unit_tests/test_word_pasting.py
@@ -1,0 +1,117 @@
+import compression_entropy
+import word_pasting as wp
+import unittest
+
+import string
+
+# TODO: replace by mocking
+MISMATCHER_PATH = '/home/pablo/ownCloud/WordOrderBibles/Literature/ThirdRound/dataverse_files/shortestmismatcher.jar'
+
+
+class TestWordPasting(unittest.TestCase):
+    def test_mask_word_structure(self):
+        verse_tokens = [['I', 'am', 'nice', 'and', 'clean', 'and', 'that', 'matters', '.']]
+        shuffled = wp.mask_word_structure(verse_tokens,
+                                          string.ascii_letters,
+                                          [1] * len(string.ascii_letters))
+        self.assertEqual(1, len(shuffled))
+        self.assertEqual(len(verse_tokens[0]), len(shuffled[0]))
+        self.assertTrue(all([verse_tokens[0][i] != shuffled[0][i] for i in range(len(shuffled[0]))]))
+        self.assertTrue(all([len(verse_tokens[0][i]) == len(shuffled[0][i]) \
+                             for i in range(len(shuffled[0]))]))
+        self.assertEqual(shuffled[0][3], shuffled[0][5])
+
+    def test_mask_word_structure_across_verses(self):
+        verse_tokens = [['This', 'is', 'it'], ['It', 'is', 'good']]
+        shuffled = wp.mask_word_structure(verse_tokens,
+                                          string.ascii_letters,
+                                          [1] * len(string.ascii_letters))
+        self.assertEqual(2, len(shuffled))
+        self.assertEqual(shuffled[0][1], shuffled[1][1])
+
+    def test_join_verses(self):
+        verse_tokens = [['I', 'hate', 'this'], ['I', 'love', 'this']]
+        text = compression_entropy.join_verses(verse_tokens, insert_spaces=True)
+        self.assertEqual('I hate this I love this', text)
+
+    def test_get_text_length(self):
+        sequences = ['Can you see the real me'.split(), 'Tommy can you hear me'.split()]
+        length = compression_entropy.get_text_length(sequences)
+        self.assertEqual(45, length)
+
+    def test_truncate(self):
+        sequences = ['Can you see the real me'.split(), 'Tommy can you hear me'.split()]
+        excedent = 7
+        truncated = compression_entropy.truncate(sequences, excedent)
+        expected = ['Can you see the real me'.split(), 'Tommy can you'.split()]
+        self.assertEqual(expected, truncated)
+        self.assertEqual(['Can you see the real me'.split(), 'Tommy can you hear me'.split()], sequences)
+
+    def test_replace_words(self):
+        verse_tokens = ['I', 'hate', 'this', '.', 'I', 'love', 'this']
+        characterized = wp.replace_words(verse_tokens)
+        self.assertEqual(len(verse_tokens), len(characterized))
+        self.assertEqual(characterized[0], characterized[4])
+        self.assertEqual(characterized[2], characterized[6])
+        self.assertEqual(5, len(set(characterized)))
+
+    def test_replace_top_bigram(self):
+        verses = ['Congratulations, you have finished installing TWiki!'.split(),
+                  'Replace this text with a description of your new TWiki site and links to content.'.split(),
+                  'To learn more about TWiki, visit the new TWiki web.'.split()]
+        replaced = wp.replace_top_bigram(verses)
+        self.assertEqual(verses[0], replaced[0])
+        expected = [verses[0],
+                    ['Replace', 'this', 'text', 'with', 'a', 'description', 'of', 'your', 'new TWiki', 'site', 'and',
+                     'links', 'to', 'content.'],
+                    ['To', 'learn', 'more', 'about', 'TWiki,', 'visit', 'the', 'new TWiki', 'web.']]
+        self.assertEqual(expected, replaced)
+        # Check that a copy was made even when no changes were made
+        expected[0] = []
+        self.assertEqual('Congratulations, you have finished installing TWiki!'.split(), replaced[0])
+
+    def test_replace_top_bigram_all_merged(self):
+        verses = [['Congratulations, you have finished installing TWiki!'],
+                  ['Replace this text with a description of your new TWiki site and links to content.'],
+                  ['To learn more about TWiki, visit the new TWiki web.']]
+        replaced = wp.replace_top_bigram(verses)
+        self.assertEqual([], replaced)
+
+    def test_merge_positions(self):
+        verses = [['I', 'love', 'the', 'nightlife', 'and', 'I', 'do', 'not', 'make', 'a', 'big', 'fuss', 'about', 'it'],
+                  ['Belgium', 'plays', 'ugly'],
+                  ['No', 'hubo', 'otro', 'como', 'Forlan']]
+        positions = [(0, 4), (0, 7), (2, 1)]
+        merged = wp.merge_positions(verses, positions)
+        expected = [['I', 'love', 'the', 'nightlife', 'and I', 'do', 'not make', 'a', 'big', 'fuss', 'about', 'it'],
+                    ['Belgium', 'plays', 'ugly'],
+                    ['No', 'hubo otro', 'como', 'Forlan']]
+        self.assertEqual(expected, merged)
+        # Check that a copy was made even when no changes were made
+        expected[1] = []
+        self.assertEqual(['Belgium', 'plays', 'ugly'], merged[1])
+
+    def test_join_words(self):
+        tokens = 'I love the nightlife and I do not make a big fuss about it'.split()
+        joined = wp.join_words(tokens, [5, 2])
+        expected = ['I', 'love', 'the nightlife', 'and', 'I do', 'not', 'make', 'a', 'big', 'fuss', 'about', 'it']
+        self.assertEqual(expected, joined)
+
+    def test_join_words_copy(self):
+        # Check that a copy was made even of the untouched verses
+        tokens = 'I love the nightlife and I do not make a big fuss about it'.split()
+        joined = wp.join_words(tokens, [])
+        expected = ['I', 'love', 'the', 'nightlife', 'and', 'I', 'do', 'not', 'make', 'a', 'big', 'fuss', 'about', 'it']
+        self.assertEqual(expected, joined)
+
+    def test_create_word_pasted_sets(self):
+        id_verses = {0: [['I', 'am', 'here', 'I', 'am']], 1: [['I am here']]}
+        steps_to_save = {0, 1}
+        word_pasted_sets = wp.create_word_pasted_sets(id_verses, steps_to_save)
+        expected = {0: {0: [['I', 'am', 'here', 'I', 'am']], 1: [['I am', 'here', 'I am']]},
+                    1: {0: [['I am here']]}}
+        self.assertEqual(expected, word_pasted_sets)
+
+
+if __name__ == "__main__":
+    unittest.main()
